+++ package.json
@@ -65,7 +65,7 @@
     "upgrade:design-system": "yarn up '@showtime-xyz/universal.*'"
   },
   "resolutions": {
-    "babel-preset-expo": "9.1.0",
+    "babel-preset-expo": "9.3.1",
     "react": "18.2.0",
     "react-dom": "18.2.0",
     "react-native-svg": "13.8.0",
