+++ packages/app/components/footer/index.tsx
@@ -63,4 +63,4 @@ const Footer = () => {
   );
 };
 
-export { Footer };
+export default Footer;
