+++ packages/app/components/header/index.ts
@@ -1,5 +1,8 @@
+import { Header } from "./header";
+
 export * from "./header";
-export * from "./header-center";
 export * from "./header-left";
 export * from "./header-right";
 export * from "./header-cancel";
+
+export default Header;
