+++ packages/app/components/settings/index.ts
@@ -1,1 +1,4 @@
+import { Settings } from "./setting";
+
 export { Settings } from "./setting";
+export default Settings;
