+++ packages/app/components/trending/index.tsx
@@ -1,1 +1,4 @@
+import { Trending } from "./trending";
+
 export { Trending } from "./trending";
+export default Trending;
