+++ packages/app/pages/profile/index.web.tsx
@@ -1,11 +1,13 @@
 import { useEffect } from "react";
 
+import dynamic from "next/dynamic";
 import { SWRConfig } from "swr";
 
 import { useRouter } from "@showtime-xyz/universal.router";
 
-import { ProfileScreen } from "app/screens/profile";
-
+const ProfileScreen = dynamic(() => import("app/screens/profile"), {
+  ssr: false,
+});
 function ProfileRouter({ fallback = {} }: { fallback?: object }) {
   const router = useRouter();
