+++ packages/app/screens/edit-profile.tsx
@@ -1,8 +1,13 @@
+import dynamic from "next/dynamic";
+
 import { withModalScreen } from "@showtime-xyz/universal.modal-screen";
 
-import { EditProfile } from "app/components/edit-profile";
 import { useUser } from "app/hooks/use-user";
 import { useTrackPageViewed } from "app/lib/analytics";
+
+const EditProfile = dynamic(() => import("app/components/edit-profile"), {
+  ssr: false,
+});
 
 export const EditProfilePage = () => {
   useTrackPageViewed({ name: "Edit Profile" });
