+++ packages/app/screens/settings.tsx
@@ -1,7 +1,11 @@
+import dynamic from "next/dynamic";
+
 import { withColorScheme } from "app/components/memo-with-theme";
-import { Settings } from "app/components/settings";
 import { useTrackPageViewed } from "app/lib/analytics";
 
+const Settings = dynamic(() => import("app/components/settings"), {
+  ssr: false,
+});
 const SettingsScreen = withColorScheme(() => {
   useTrackPageViewed({ name: "Settings" });
