+++ packages/app/screens/trending.tsx
@@ -1,6 +1,11 @@
+import dynamic from "next/dynamic";
+
 import { withColorScheme } from "app/components/memo-with-theme";
-import { Trending } from "app/components/trending";
 import { useTrackPageViewed } from "app/lib/analytics";
+
+const Trending = dynamic(() => import("app/components/trending"), {
+  ssr: false,
+});
 
 const TrendingScreen = withColorScheme(() => {
   useTrackPageViewed({ name: "Trending" });
