+++ packages/design-system/hidden/index.tsx
@@ -1,4 +1,4 @@
-import { useMemo } from "react";
+import { useMemo, useState, useEffect } from "react";
 import { Platform, useWindowDimensions } from "react-native";
 
 import {
@@ -16,12 +16,17 @@ type HiddenProps = {
 
 export const Hidden = (props: HiddenProps) => {
   const { width } = useWindowDimensions();
+  const [mounted, setMonted] = useState(false);
   const { from, platform, until, children } = props;
+
   const currentBreakpoint = useMemo(
     () => sortedBreakpointKeys.find((key) => width >= breakpoints[key]),
     [width]
   );
-
+  useEffect(() => {
+    setMonted(true);
+  }, []);
+  if (!mounted) return null;
   if (!from && !until && !platform) {
     return null;
   }
