+++ packages/design-system/label/index.web.tsx
@@ -1,9 +1,3 @@
 import { Text, Props as TextProps } from "@showtime-xyz/universal.text";
 
-export const Label = (props: TextProps) => (
-  <Text
-    {...props}
-    // @ts-ignore
-    accessibilityRole="label"
-  />
-);
+export const Label = (props: TextProps) => <Text {...props} />;
