+++ packages/design-system/pressable-hover/index.tsx
@@ -1,4 +1,4 @@
-import { useMemo } from "react";
+import { useMemo, useState, useEffect } from "react";
 import {
   Pressable as ReactNativePressable,
   PressableProps,
@@ -14,6 +14,7 @@ export type Props = Omit<PressableProps, "tw"> & {
 const StyledPressable = styled(ReactNativePressable);
 
 export function PressableHover({ tw, ...props }: Props) {
+  const [mounted, setMounted] = useState(false);
   const twWithHover = useMemo(
     () =>
       [
@@ -22,6 +23,10 @@ export function PressableHover({ tw, ...props }: Props) {
       ].join(" "),
     [tw]
   );
+  useEffect(() => {
+    setMounted(true);
+  }, []);
+  if (!mounted) return null;
 
   return <StyledPressable {...props} tw={twWithHover} />;
 }
