+++ workspace.jsonc
@@ -54,7 +54,7 @@
         "@types/react-dom": "^18.0.10",
         "@types/react-native": "^0.71.1",
         "@types/testing-library__jest-dom": "5.14.5",
-        "babel-preset-expo": "^9.2.2",
+        "babel-preset-expo": "^9.3.1",
         "blurhash": "2.0.4",
         "core-js": "^3.27.2",
         "expo-blur": "12.0.1",
